autohi,refactor: use auto-highlights lib and refactor location
this commit utilizes the "litee.lib.highlights.auto" library to
automatically highlight source code lines when the cursor resides on a
call tree node.

additionally, the use of the "resolve_location" method is removed and
the location information along with references to symbols are now placed
on the top level of the node table.

this conforms to the latest changes in "litee.lib" which deprecate this
method.

Signed-off-by: ldelossa <louis.delos@gmail.com>