useability: use extmarks to keep jumps in sync
this pull uses extmarks to make jump highlights use similar to how marks
work in Neovim.

now, an extmark is created for every calltree node and their references.

an auto command is added which syncs up the calltree node's "location"
field with the current position of the extmark.

this works pretty well and is more light weight then building a new
calltree everytime.

there is a limitation, like Neovim marks, copy and pasting the area of
text whch defines the Symbol will not move the extmark when pasted.