useability: make references work correctly.
this commit changes how calltree.nvim handles references to symbols and
how calltree works fundamentally to some degree.

now, when browsing the calltree, the symbol at which the cursor is ontop
of will be navigated too.

if there is more then one reference for the symbol under the cursor, for
example, a particular function is called more then once from the root
symbol, then hitting (by default) "n" will navigate to the next
reference.